requirements.txt
# Dependencias principales
streamlit==1.24.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0
wordcloud==1.9.2
networkx==3.2.1
kaleido==0.2.1

# NLP y procesamiento de texto
nltk==3.8.1
spacy==3.5.3
python-Levenshtein==0.21.1
sentence-transformers==2.2.2
gensim==4.3.1
bibtexparser==1.4.0

# Exportación y reportes
reportlab==4.0.4

# Para desarrollo
pytest==7.4.0
black==23.3.0
mypy==1.4.1